refactor(tls): share material validation boundary
