+++ src/config/validation/config_validators.rs
@@ -188,7 +188,7 @@ impl Validate for ServerConfig {
         if let Some(tls) = &self.tls {
             tls.validate()?;
             #[cfg(feature = "gateway")]
-            crate::server::tls::validate_rustls_config(tls)?;
+            crate::utils::tls::validate_rustls_config(tls)?;
         }
 
         Ok(())
